mcp-redis/requirements.txt
azure-functions
redis
numpy
openai
python-dotenv